pyproject.toml
[project]
name = "photobooth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "gradio>=6.0.2",
    "matplotlib>=3.10.7",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "pilgram==2.0.0",
    "pillow==12.0.0",
    "python-multipart>=0.0.20",
    "pyyaml==6.0.3",
    "qrcode==8.2",
    "uvicorn>=0.38.0",
]
